Update pizza ordering interface and add order removal feature
In the pizza ordering interface, binding was added to the pizza size input to provide real-time interactivity. Unnecessary nesting of the cancel and order buttons was removed for code cleanup. Also, implemented a feature to remove pizzas from the order and added visibility of the 'Your Order' side bar dependent on if any pizzas have been added to the order yet.